# docs/rapidapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lottizen Data API
  version: "1.0.0"
  description: >-
    Lottery data for Canada, the US and Europe: winning numbers, 40+ years of
    draw history, hot/cold/overdue number statistics, and the only Ontario
    scratch-ticket remaining-prize tracker in Canada. Every dataset refreshes
    daily from official lottery-operator sources. Japan coverage is coming
    soon.
  contact:
    name: Lottizen
    url: https://lottizen.com/api
  license:
    name: Proprietary — see https://lottizen.com/methodology
servers:
  - url: https://lottizen.com/api/v1
    description: Production (subscribe via RapidAPI; requests are proxied and billed there)
...
paths:
...
  /games/{slug}:
    get:
      operationId: getGame
      summary: Get a single game
      tags: [Games]
      parameters:
        - $ref: "#/components/parameters/GameSlug"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Game" }
                  meta: { nullable: true }
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    GameSlug:
      name: slug
      in: path
      required: true
      description: Game slug, as returned by GET /games.
      schema: { type: string, example: lotto-max }
  responses:
    NotFound:
      description: No matching game found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: GAME_NOT_FOUND
              message: "No live game found for slug 'foo'."
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: GAME_NOT_FOUND }
            message: { type: string }
    Game:
      type: object
      properties:
        slug: { type: string, example: lotto-max }
        name: { type: string, example: Lotto Max }
        country: { type: string, enum: [CA, US, EU] }
        countryName: { type: string, example: Canada }
        agency: { type: string, example: National }
        operator: { type: string, example: the Interprovincial Lottery Corporation }
        region: { type: string, example: Canada-wide }
        format: { type: string, enum: [lotto, keno, digit, special] }
        pick: { type: integer, example: 7 }
        max: { type: integer, example: 52 }
        hasBonus: { type: boolean }
        bonusLabel: { type: string, nullable: true, example: Bonus }
        bonusMax: { type: integer, nullable: true, example: 52 }
        bonusCount: { type: integer, example: 1 }
        drawDays:
          type: array
          items: { type: string }
          example: [Tuesday, Friday]
        price: { type: number, example: 6 }
        currency: { type: string, enum: [CAD, USD, EUR, GBP] }
        blurb: { type: string }
        progressive: { type: boolean }
        drawCount: { type: integer, example: 753 }
        dataSince: { type: string, format: date, nullable: true }
...